@@ -4,12 +4,12 @@ plugins {
 
 android {
     namespace 'com.zhoujh.lvtu'
-    compileSdk 33
+    compileSdk 34
 
     defaultConfig {
         applicationId "com.zhoujh.lvtu"
         minSdk 27
-        targetSdk 33
+        targetSdk 34
         versionCode 1
         versionName "1.0"
 
@@ -30,9 +30,10 @@ android {
 
 dependencies {
 
-    implementation 'androidx.appcompat:appcompat:1.4.1'
-    implementation 'com.google.android.material:material:1.5.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
+    implementation 'androidx.appcompat:appcompat:1.7.0'
+    implementation 'com.google.android.material:material:1.12.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    implementation 'androidx.activity:activity:1.9.3'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
@@ -58,6 +59,7 @@ dependencies {
     // glide
     implementation 'com.github.bumptech.glide:glide:4.16.0'
 
-//    implementation 'io.github.FlyJingFish.OpenImage:OpenImageFullLib:2.1.9'
+    //OpenImageFullLib 默认已经包含了OpenImageGlideLib
+    implementation 'io.github.FlyJingFish.OpenImage:OpenImageFullLib:2.3.6'
     implementation 'jp.wasabeef:glide-transformations:4.3.0'
 }
