@@ -1,6 +1,6 @@
 package com.zhoujh.lvtu.model;
 
-import java.util.Date;
+import java.time.LocalDateTime;
 
 public class User {
     private String userId;       // 用户ID (UUID)
@@ -9,17 +9,17 @@ public class User {
     private String email;        // 邮箱
     private String password;     // 密码
     private Integer status;      // 用户状态
-    private Date createTime;     // 创建时间
+    private LocalDateTime createTime;     // 创建时间
     private Integer gender;      // 性别
     private Integer age;         // 年龄
     private String birth;        // 生日
     private String avatarUrl;    // 头像路径
-    private Date updateTime;
+    private LocalDateTime updateTime;
 
     public User() {
     }
 
-    public User(String userId, String userName, String phoneNum, String email, String password, Integer status, Date createTime, Integer gender, Integer age, String birth, String avatarUrl,Date updateTime) {
+    public User(String userId, String userName, String phoneNum, String email, String password, Integer status, LocalDateTime createTime, Integer gender, Integer age, String birth, String avatarUrl,LocalDateTime updateTime) {
         this.userId = userId;
         this.userName = userName;
         this.phoneNum = phoneNum;
@@ -82,11 +82,11 @@ public class User {
         this.status = status;
     }
 
-    public Date getCreateTime() {
+    public LocalDateTime getCreateTime() {
         return createTime;
     }
 
-    public void setCreateTime(Date createTime) {
+    public void setCreateTime(LocalDateTime createTime) {
         this.createTime = createTime;
     }
 
@@ -122,11 +122,11 @@ public class User {
         this.avatarUrl = avatarUrl;
     }
 
-    public Date getUpdateTime() {
+    public LocalDateTime getUpdateTime() {
         return updateTime;
     }
 
-    public void setUpdateTime(Date updateTime) {
+    public void setUpdateTime(LocalDateTime updateTime) {
         this.updateTime = updateTime;
     }
 }
