@@ -1,5 +1,6 @@
 package com.zhoujh.lvtu;
 
+import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
@@ -8,10 +9,23 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.zhoujh.lvtu.find.AddPostActivity;
+
 public class FindFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_find, container, false);
+        View view = inflater.inflate(R.layout.fragment_find, container, false);
+
+        initView(view);
+
+        return view;
+    }
+
+    private void initView(View view) {
+        view.findViewById(R.id.to_add_post).setOnClickListener(v -> {
+            Intent intent = new Intent(getActivity(), AddPostActivity.class);
+            startActivity(intent);
+        });
     }
 }
